requirements.txt
yandex_music
python-dotenv
wxPython
pillow
pypresence
selenium==4.8.1
requests
webdriver_manager